// backend/src/types/query_types.ts
// Types for stuff coming from the frontend

// Query generated by the front end's Query Editor, after being handled by the hsfl-cosmos-datasource backend
export interface QueryType {
    // From the front end, JSON string that unmarshals into a QueryObject
    query: queryString
    // MJD time range start time
    from: number;
    // MJD time range end time
    to: number;
}
type queryString = string;
export interface QueryObject {
    // Query type
    type: string;
    // Query argument
    arg: string;
    // Return only the latest result
    latestOnly: boolean;
    // Various filterings
    filters: QueryFilter[];
    // Functions to apply to results for additional results
    functions: QueryFunction[];
}
export type FilterType = 'node' | 'name' | 'col';
export type CompareType = 'equals' | 'contains';
export type FunctionType = 'sum' | 'magnitude';
export interface QueryFilter {
    filterType: FilterType;
    compareType: CompareType;
    filterValue: string;
}
export interface QueryFunction {
    functionType: FunctionType;
    args: string[];
}